Add FedDecision Class and Add Support for AggBinary FED Hop
This is the initial step for supporting compilation of FED instructions. The constructLops() method needs to be adapted to handle FED type hops for the other implementations of Hop similar to what is done in this commit for AggBinaryOp.

Edit HOP ExecType and Output Federation

This commit edits the way ExecType is updated and adds a federatedOutput field in HOPs which can be used later to define a sequence of federated instructions with no GET calls between the federated instruction calls

Remove FedDecision Class

Add Basic Propagation of Federated Output Flag

This is only propagated for a single use case which involves Binary, AggBinary, and Transpose.

Remove Empty FedDecision File